+++ tests/__main__.py
@@ -2,7 +2,7 @@
 Unit tests for the introc module
 
 :author:  Walker M. White (wmw2)
-:version: July 24, 2018
+:version: June 9, 2019
 """
 import unittest
 
@@ -11,13 +11,13 @@ def suite():
     """
     Creates the test suite for all packages except turtle (which is graphical)
     """
-    modules  = ( 'test_testcase','test_strings','test_tuples','test_colors','test_geom','test_filetools','test_urltools')
+    modules  = ( 'test_testcase','test_strings','test_tuples','test_colors',
+                 'test_geom','test_filetools','test_urltools','test_modlib')
     alltests = unittest.TestSuite()
-    for module in map(__import__, modules[:-1]):
+    for module in map(__import__, modules):
         alltests.addTest(unittest.findTestCases(module))
     return alltests
 
 
 if __name__ == '__main__':
     unittest.main(defaultTest='suite')
-