+++ tests/files/module1.py
@@ -1,0 +1,10 @@
+"""
+Simple test module for modlib
+
+:author:  Walker M. White (wmw2)
+:version: June 9, 2019
+"""
+
+
+x = 5
+y = 'Hello'